Refactor: Make USNY base class, USGS inherits and adds Good Friday
- Move all SIFMA holiday logic from _sifma.py to USNewYork class
- USNewYork is now the simpler base (no Good Friday)
- USGovernmentSecurities inherits from USNewYork and adds Good Friday
- Delete _sifma.py as it's no longer needed
- This follows the principle of building from simpler to more complex

Addresses KJhellico's architectural feedback
